requirements.txt
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
google==3.0.0
google-api-core==1.27.0
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
grpcio==1.37.1
httplib2==0.19.1
idna==2.10
numpy==1.19.5
packaging==20.9
Pillow==8.2.0
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
uritemplate==3.0.1
urllib3==1.26.4
